THE BIG RE-ORGANISATION!
All the sources from src/tracer/prelude have been merged in with
the normal sources in src/prelude.  There are new #ifdef's for
those parts which differ.

At the same time, we have simplified the prelude Makefile build system,
introducing a single Makefile.common which holds the rules for both
normal and tracing builds, and replacing the individual Makefiles in
each library directory with a simple call to the common Makefile.

In addition, the machine-specific parts of the prelude (mainly
DErrNo.hs) have been separated out so that several machines can
build simultaneously from the same source tree.