Foreign importing a C function :: () -> res was rather nasty because
it tried to pluck a value of void type from Haskell and give it to C -
also the actual C call then tried to use the value!  I've done a small
hack so that a function of arity=1 with arg () goes through fine.  I
can't see anyone needing to pass a () arg for any reason other than
to force re-evaluation on multiple calls, so that should be ok I hope.