Removed bug that tracing did not handle newtype.
The problem was that the compiler replaces the data constructor
defined by a newtype by the function id (if the data constructor
has no arguments, which is always the case after the tracing transformation).
This function id was however the transformed function, that is, it was no
longer the simple identity.
To solve the problem I introduced a new function _id in the Prelude
which is not transformed and can hence be used to replace the
data constructor.
Also added type declarations and comments.