Small changes towards getting traced-numerics to work.  In a Fractional
context where the type is unknown, we can't resolve a literal to either
Float or Double at compile-time, so we have to generate `fromRational l'
where the literal l is a constructed Rational.  Unfortunately,
traced Rationals have a pretty complex structure, so we have to augment
the auxiliary `conRational' with extra arguments in order to build
the traced literal Rational correctly.  It still doesn't work though.  :-(