The -ansiC backend had a huge plateau space-leak.  By splitting
code-generation into two passes (labels then code), we greatly
reduce the leak, but introduce a shorter spike later.  Change the
representation of the code from composed 'shows' to concat'ed strings,
and the spike is halved (but still spiky).  Overall, much better space
usage (although it is still not as good as it could be).  Time usage
appears to be exactly the same as before (although still not as good
as with the -noansiC backend).