Removed #if defined(TRACING) from src/prelude/*/*.hs where it was used before,
because the tracing transformation couldn't cope with type synonyms.

Discovered a bug in the tracing transformation of type synonyms and
removed it.
The function unwrapNT needs a special expand function, the one in
TypeUnify is not suitable because it uses the idempotent closure of the
substitution.

Also added some type declarations and comments.