Replaced the old ugly and inefficient pretty printer by a new efficient one.
Hence modules PPLib, PPSyntax and StrSyntax have been removed.
They are replaced by PrettyLib and PrettySyntax.
There are some new compiler options to set output width, indentation of nesting and qualification of identifier.
For printing NewType and NT still the module Nice is used, that will have to be changed in the future.