Add a new primitive implementation of hPutStr that packs Strings
before outputting them.  We 'chunk' the string as we evaluate it
lazily, so as not to use too much space if the string is very long.
But we also need to be careful to do this correctly with respect
to the buffering mode (char,line,block).

Initial performance measurements show a 5-10x speedup for output,
depending on the length of the original string and the buffering
mode.