Tracing does not work with this version.
Changed the trace transformation in DbgTrans, the combinators in DebugPrelude and a few minor things in the compiler to call functions that create the trace on file (mkTAp1,..) instead of the old algebraic data types. Actually defined the new functions (mkTAp1,..) in terms of the old algebraic data types, so that tracing still partially works. However, trusting is always off and updating of Sats in the old way cannot be done with the new interface.