Correct bug in desugaring of do notation in DbgDataTrans.
It created DeclSccs instead of DeclParse.


Changed definition of combinator tif in DebugPrelude, so that
an unevaluated if is shown in unevaluated form instead of
always being evaluated.
The following used to give a blackhole before:
main = print (True || if x then False else False)
x :: Bool
x = x


Also performed some inlining in combinators of DebugPrelude,
but effect on runtime is not noticeable.