Fixed definition of instance Enum for tracing.
For tracing the instance lacked definitions for toEnum and fromEnum.
The same definitions as for untraced programs can be used.
Hopefully not more instance definitions are lacking for tracing;
here it really hurts that the compiler, following the Haskell standard,
cannot detect missing instance definitions.

Unfortunately a missing instance definition leads to faulty byte code
for tracing in the current compiler. Still have to modify transformation
to give proper runtime error.