Fixed transformation of default methods.
A missing default method lead to a corrupted trace file and sometimes even a segmentation fault. The problem was that the call to `fatal' was not wrapped within an R. Furthermore the error message "No match in pattern." was rather misleading.

Now transform default method like any other function definition. The error message which has been introduced by the renamer is transformed like a normal expression. Changed renamer so that this compiler introduced expression has no source positions. So the explicit call of error is visible in the trace, but the browser user cannot ask for the source position.