Introduced trusted and suspected variant of application combinators.
The suspected application combinators ap and rap always create a trace node for an application.
The trusted application combinators tap and trap create a trace node for an application, if the function is suspected (no need to trust parent).

Changed transformation to use the new application combinators.

Trace file unchanged.
Traced code 5-10% faster on small heaps, unchanged for large heaps.