Record trusted partial application of partially applied constructors.
If in a trusted application the function was a partial application of a data constructor, then this trusted application was not recorded in the trace.
E.g.  map ((,) 1) [2,3] showed only hidden nodes as elements of the result list. Changed fileformat.c so that data constructors are always untrusted (independent of their definition module). Then the aforementioned applications are recorded in the trace.

Also changed PreludeDebug.hs so that oversaturated applications in trusted code are only recorded in the trace if the result of the saturated application is untrusted.

Trace files become slightly (~1%) larger than before without noticeable effect on speed.