Reintroduced Sats
Now in suspected code an application or Caf name only has no Sat, if its result is the same as the result of the parent.
Note that in trusted code there are less Sats. A Sat is only constructed, if the unevaluated expression can be returned to suspected code (especially as argument of a constructor). Actually this is not quite true and needs to be corrected.

There are now more SatCs than before. Filesize grows by 1%, speed change not noticeable.

Undid part of previous implementation of field name selectors. They disturbed correct working of Cafs. So for now only selectors for data types that are defined in the same module will work.