Introduced even more Sats and resurrected field selectors.
The application of a saturated application never had a Sat. Now it has.
If a trusted application is recorded (because the function is suspected), then it has a Sat.

There are now even more SatCs. Filesize grows by ~2%, runtime grows by ~2%.

Corrected implementation of transformed field selectors. The definitions are derived at transformation time and subsequently transformed. Because nhc derives the type of imported selectors from data definitions, these types have to be patched by the transformation.

Now application of a field selector to a value that does not contain the field is caught correctly in the transformed code.