Completed tracing of named fields.
Now tracing of record construction and update is supported as well.
However, the trace only shows the evaluation of the desugared program,
that is, record construction is shown as data constructor application
(possibly with `undefined' arguments) and record update is shown as
use of the `case' construct.

The trace file format needs to be extended to enable showing this operations properly. It would be nice if for all data constructors which have named fields these were always shown in the browser.

Also:
Corrected bug in records: x(C{x=1}{x=2}) evaluated to 1 instead of 2.
Added missing type declaration of `readFile' to Prelude.T.hi.
Handle the definition of `undefined' specially in the tracing transformation.
`undefined' is treated as nullary function, not as caf. Thus it has a proper parent.
Corrected combinators for nullary functions. They were too strict.