Corrected typo in definition of combinator fun6.
The last line was missing in the definition. Unfortunately the definition was type correct, only with a different type. Because Prelude.T.hi is handmade, the inconsistency could not be detected by the system.
It seems the examples Adjoxo,Mate etc. don't have a function with six arguments, otherwise they would surely have core dumped.