Corrected bug of unevaluated partial applications of constructors.
The combinator `eagerSat' was not the source of the bug.
Hence I resurrected it.

The source was that data constructors defined in trusted modules were trusted. However, they should always be suspected so that their application is recorded in the trace. Changed fileformat.c accordingly.

The Mate trace file grows by 2%.