Improved trace construction for suspected pattern bindings.
Introduced an indirection in transformation of suspected pattern bindings, so that the whole rhs is reachable in hat-trail. However, the connection to the names of the variables are lost, when they are evaluated. There seems to be no natural point where to add them to the trace.

Increases Mate trace file by 1%.

Corrected bug in hat-trail which made it impossible to see parents of indirections in many cases. However, hat-trail still does not deal with indirections completely correctly. Pointing at an expression surrounded by an indirection highlights the parent of the expression in yellow, not the parent the indirection points to (and which is shown if asked for).

First thought that hat-connect doesn't deal with indirections correctly, but it is in fact correct.