Add variable names to top-level trusted pattern bindings.
Similar to top-level trusted cafs. In practise this probably hardly occurs. However, this variable names are only shown if the variables are not evaluated. If they are, they are no longer reachable in the trace. That may mean, that there is only root as printable representation of the result. However, it is unclear where the variable names should be added to the trace.