Introduce Sat for outermost application on rhs of pattern binding.
The outermost application on a rhs usually doesn't have a Sat, because the result equals the result of the parent, which is the lhs. However, for pattern bindings the lhs is not the parent. So hat-observe needs the Sat in that case.
If the rhs of a pattern binding is a single lambda-bound variable, we no longer create an indirection. However, this indirection seems pretty pointless anyway, the parent of the variable value is probably more interesting.

Increases Mate trace file by 0.02%.