Introduced specially marked lonely Sats
Sats that don't come with a preceding application or name have the bit with value 8 set. This shall enable hat-detect and hat-observe to show reductions correctly and not be confused by applications/names without Sats that a followed by an unrelated Sat by chance.
Small changes in many files. The SatB stack and the SatC queue need additional entry for the sort of a caf (lonely or not).