Corrected several bugs in hat-stack.
Set string buffers for expressions in readTraceAt and infixPrint to 10000 instead of 255 (no final solution).
If an expression has no source reference hatStack prints a message that says so instead of seg faulting.
Hidden itself is never a stack entry. If a parent is a hidden, we follow the hidden parent. Hidden is only shown in subexpressions.
Infix printing is only used when the operator has at least 2 arguments.

Improvements:
Expressions are cut off at a depth of 6 to avoid very large expressions.
if, case and guard redexes are shown correctly, with the parent as next stack entry.