Permit hat-trail to display parent-colouring correctly.
To achieve this, we add an extra field to most of the messages sent
from hat-connect to hat-trail.  This field contains a prediction
(obtained via the new routine peekTrace()) of the node number that
will eventually be the /displayed/ parent for a particular node,
i.e. by skipping over Hiddens and Sats and following the left pointer
of Indirections.