Corrected tracing of irrefutable patterns.
Extended tracing to handle pattern bindings with an as-pattern at the top, e.g. x@(_,y) = ....
However, note that tracing of pattern bindings is still more a hack than a proper solution. Unclear what proper solution should be. Tracing of pattern bindings is a problem at a level similar to tracing list comprehensions or do notation.