Complete library src/prelude/Hat for portable Hat.
The library now contains a minimal prelude wrapper.
There is a Makefile which works, but produces many warnings.

The compiler option -hx has been replaced by -hat.
Envoking it transforms a module `Name' into a tracing module `TName'.
If the module is `Main' the name is unchanged and a special function
definition of `main' is added (the filename is still prefixed by `T').

The transformation works for an extended hello world program.
For some reason (packString not always working?) the module name `Main' is followed by many bytes of rubish before the final \0 in the trace file.
hat-trail works fine with the trace.
hat-detect shows some strange reductions and hat-observe doesn't show reductions of all identifiers. This is with the v1.08 tools, because the transformation doesn't yet produce any toplevel identifiers (all are in old format and hence local).

Changed some includes of hat.h to art.h (forgotten last time).