Basically finished implementation of portable transformation and combinators for untraced code. However, it has not yet been tested!
The combinators hide and hidden are implemented an seem to be correct.

Also corrected different handling of cafs and functions with arguments: Methods in class and instance definitions always have to be handled as functions with arguments, because the arity may vary for each instance. Declarations of types for cafs are transformed differently from declarations of types for functions with arguments.