Implemented "deriving Eq" in portable transformation.
The general mechanism and the specific one for the class Eq are implemented.
Deriving for the remaining classes still needs to be done.

The derived code is untraced.

I use a hack for determining the context of an instance. It is the context that will "commonly" be needed, but it is not correct in all cases. A correct implementation would require much more effort: context reduction (with knowledge about all classes, instances and type synonyms) and a fixed point iteration.