Corrected two bugs in new handling of newtype data constructors and twiddle in patterns.
DeclsScc can occur because of desugaring of record expressions.

Variations of the variable "irrefutableDummy" need to be created that don't clash with each other. First, the variable need to be qualified to avoid clashes between modules. Second, the source position is part of the name to avoid clashes within the Prelude. The later is only a hack; there may still be a clash but it hasn't been noticed yet.