Modifyied both nhc98 runtime system and hat library to catch runtime errors.
The nhc98 runtime system now has the global variable
void (*haskellErrorHandler)(char *errorMsg) = NULL;
If it is not null, it will be called for runtime errors that are not handled by Haskell's "error"; i.e. currently "out of heap" and "black hole".

The hat_open function sets haskellErrorHandler to an appropriate function that writes the error message to the trace and closes it. (only if compiled for nhc98)

Also added handling of arithmetic error signals to the hat library.