Fixed bug in handling of pattern bindings.
Pattern bindings with newtype data constructors where incorrectly transformed; AuxLabelAst already added bogus definitions for the variables within these data constructors, because it treated these variables to have only a local scope within the pattern definition.

Also added error message if hat-trans is applied to a program with (..) after a type or class in an export or import list.