Better handling of package libraries.  Previously, the package
import directories were detected at installation time, so the
addition of a new package required a re-configure of hmake.
Also, because all of the package dirs were searched on every
invocation, hmake could not warn of a missing `package' flag.
Now, package dirs are detected at runtime, and only for the
requested packages - this fixes both problems.