Add instances of Data.Bits.Bits for Int/Word[8,16,32,64] in a different
way.  The module Data.Bits is only available from the base package,
which is normally added to the link line *before* the standard libraries,
hence if the instances are in the latter, you get link errors.  So move
the instances into the base package where they belong, in NHC.SizedTypes.