OK, this seems to be the end of nhc98's cunning memory layout: gcc 3.3.3
legitimately reshuffles the arrangement of the arrays (alphabetically?), so we
must not make any assumptions about their order. This is probably only the
beginning of the whole story...

NOTE: We should really abstract from the magic constants 10 (for ints) and 1
(for chars), but I was too lazy to do that now.