Don't compile code which depends on adventurous storage layout assumptions with
optimizations, otherwise gcc will legitimately move the code around.

NOTE: This is really only a very fragile workaround and may break again anytime
soon. Either nhc98's bytecode interpreter should be changed to not rely on the
assumption that certain blocks are adjacent (which will probably impose some
performance hit), or the mangler should be made more robust. In the latter case,
sed will probably not be enough, a strategy would be: Mark the different parts
of the bytecode somehow and re-arrange the assembler code in a suitable way. =>
Perl!  :-P  Somehow nhc98's mangler and the one from GHC seem to converge...

Anyway, with all these changes in the last few days, nhc98 can successfully be
bootstrapped to a stage-3 compiler again: GHC => nhc98 => nhc98