Made the type for source code positions in Extra, Pos, an abstract data type.
This required changes of many type annotations where Pos and Int were confused.
Also added some more type annotations.
In Gcode the internal representation of Pos still appears.

This is all the first step towards an extended Pos that contains line and column of both start (currently) and end (to do) of a syntactic construct.