Fix lexical error (masquerading as a parse error) with an underscore
pattern e.g. "f _2a", when the -underscore flag is off (the default).
Because the character following the underscore is neither lower nor
upper, nhc98's special treatment (ignore the underscore in deciding
whether it is a conid or varid) does not apply.  Thus, drop through to
the standard H'98 behaviour of underscore=lowercase, therefore a varid.