New utility: cabal-parse
This simple command-line utility is intended to enable the easy
retrieval of information from a .cabal file, e.g. for use in Makefiles
or shell-scripts.
    Usage: cabal-parse [-slash] fieldname filename
Where a field of the .cabal file contains a list, cabal-parse ensures
(for instance) that interspersed commas are removed for simpler
onward processing.  The optional -slash flag converts dots in module
names to slashes, to help find source files.